% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xslt.R
\name{read_xslt}
\alias{read_xslt}
\title{Read in an XSLT document}
\usage{
read_xslt(xslt_src)
}
\arguments{
\item{xslt_src}{an XSLT document (either a string with XML/XSLT, a filename
or URL pointing to a file wiht XML/XSLT)}
}
\value{
This has the same return value as \code{xml2::read_xml} but adds
       \code{xslt_document} to the class list for potential future use with
       this package.
}
\description{
Lightweight wrapper for read_xml
}
\examples{
library(xslt)

xslt_src <- '<xsl:stylesheet version="1.0"
  xmlns:xsl="http://www.w3.org/1999/XSL/Transform">

  <xsl:template match="/">
    <article>
      <title>Hello World</title>
    </article>
  </xsl:template>
</xsl:stylesheet>'

xsl <- read_xslt(xslt_src)
(class(xsl))
}

