% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xslt.R
\name{xslt_transform}
\alias{xslt_transform}
\title{Transform an XML document with an XSLT stylesheet}
\usage{
xslt_transform(xml_doc, xslt_doc, is_html = FALSE, fix_ns = FALSE)
}
\arguments{
\item{xml_doc}{an XML or HTML document (either a string with XML/HTML, a filename
or URL pointing to a file wiht XML/HTML). This can also be the result
of a value returned from \code{xml2::read_xml} or \code{xml2::read_html}, thus
enabling it to be used in pipelines.}

\item{xslt_doc}{an XSLT document (either a string with XML/XSLT, a filename
or URL pointing to a file wiht XML/XSLT). This can also be the result
of a value returned from \code{xml2::read_xml} or \code{xml2::read_html}.}

\item{is_html}{is \code{xml_doc} really an HTML? If so, set this (default \code{FALSE})}

\item{fix_ns}{intercourse the namespaces (highly useful for HTML processing - default \code{FALSE})}
}
\value{
Either objects comparable to \code{xml2::read_xml} or \code{xml2::read_html} (if the
        XSLT output method was \code{xml} or \code{html} respectively) or a single-element
        character vector with the transformed document text.
}
\description{
Transform an XML document with an XSLT stylesheet
}
\examples{
library(xml2)
library(xslt)

xml_src <- "<test/>"
xslt_src <- '<xsl:stylesheet version="1.0"
  xmlns:xsl="http://www.w3.org/1999/XSL/Transform">

  <xsl:template match="/">
    <article>
      <title>Hello World</title>
    </article>
  </xsl:template>
</xsl:stylesheet>'

doc <- read_xml(xml_src)
xsl <- read_xslt(xslt_src)

res <- xslt_transform(doc, xsl)
cat(as.character(res))

res <- xslt_transform(xslt_src, xslt_src)
cat(as.character(res))
}

